# Remove index_put from MM embeddings merging