Summary:
Previously, _merge_multimodal_embeddings used `inputs_embeds[is_multimodal]=flattened` to merge the MM embeddings. This index_put operations calls non_zero in pytorch, which forces an additional D2H shape check sync point.
This diff uses masked_scatter_ to bypass the D2H point. The latency changes from 35ms to 0.08ms.

Test Plan:
E2E Results:
Before:
```
QPS:                 0.91
Avg latency:         8.542s
Avg TTFT (client):   4186.02ms
P50 TTFT (client):   4589.01ms
P99 TTFT (client):   6897.92ms
Avg TTIT (client):   21.78ms
P50 TTIT (client):   19.99ms
P99 TTIT (client):   41.25ms
Avg TTFT (server):   5657.19ms
Avg TTIT (server):   130.71ms
Avg prefill len:     22284.20 tokens
P50 prefill len:     22284.00 tokens
P99 prefill len:     22291.00 tokens
Avg decode len:      200.00 tokens
P50 decode len:      200.00 tokens
P99 decode len:      200.00 tokens
```
After:
```
QPS:                 0.94
Avg latency:         8.456s
Avg TTFT (client):   4089.09ms
P50 TTFT (client):   4510.47ms
P99 TTFT (client):   6902.79ms
Avg TTIT (client):   21.83ms
P50 TTIT (client):   19.89ms
P99 TTIT (client):   41.02ms
Avg TTFT (server):   3808.06ms
Avg TTIT (server):   79.76ms
Avg prefill len:     22284.26 tokens
P50 prefill len:     22284.00 tokens
P99 prefill len:     22291.00 tokens
Avg decode len:      200.00 tokens
P50 decode len:      200.00 tokens
P99 decode len:      200.00 tokens
```

Rollback Plan:

Differential Revision: D79405697